+++ code/Meerkat.Security/Security/Activities/Configuration/ClaimElementCollection.cs
@@ -1,0 +1,12 @@
+﻿using Meerkat.Configuration;
+
+namespace Meerkat.Security.Activities.Configuration
+{
+    public class ClaimElementCollection : NamedConfigElementCollection<ClaimElement>
+    {
+        protected override string ElementName
+        {
+            get { return "claim"; }
+        }
+    }
+}