+++ code/Meerkat.Security/Security/Activities/Configuration/PermissionElement.cs
@@ -1,9 +1,12 @@
-﻿namespace Meerkat.Security.Activities.Configuration
-{
-    using System.Configuration;
+﻿using System.Configuration;
 
+namespace Meerkat.Security.Activities.Configuration
+{
     public class PermissionElement : ConfigurationElement
     {
+        /// <summary>
+        /// Gets or sets the roles
+        /// </summary>
         [ConfigurationProperty("roles", IsRequired = false, IsKey = false)]
         public string Roles
         {
@@ -11,11 +14,24 @@
             set { this["roles"] = value; }
         }
 
+        /// <summary>
+        /// Gets or sets the users
+        /// </summary>
         [ConfigurationProperty("users", IsRequired = false, IsKey = false)]
         public string Users
         {
             get { return this["users"] as string; }
             set { this["users"] = value; }
+        }
+
+        /// <summary>
+        /// Gets or sets the claims
+        /// </summary>
+        [ConfigurationProperty("", IsDefaultCollection = true, IsKey = false, IsRequired = true)]
+        public ClaimElementCollection Claims
+        {
+            get { return (ClaimElementCollection)this[""]; }
+            set { this[""] = value; }
         }
     }
 }