+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 1.1.0 (2016-05-24)
+* Introduce claims based authorization, permission can support multiple claims types/claims
+* Breaking change, removed Role from AuthorizationReason and replaced with Reason which records the User, Role or Claim that satisfies a permission
+
 #### 1.0.2 (2016-05-23)
 * Renamed Authorise -> Authorise throughout for consistency
 * Fix package reference, should be Common.Logging not Common.Logging.Core